pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ianuacare"
version = "0.1.0"
description = "Healthcare-oriented data pipeline and AI inference framework for Ianua products"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Ianua" }]
keywords = ["healthcare", "pipeline", "ai", "inference"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]

dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.4",
    "ruff>=0.6",
    "mypy>=1.11",
]
aws = ["boto3>=1.34", "python-jose[cryptography]>=3.3"]
postgres = ["psycopg[binary]>=3.1"]
together = ["together>=1.0"]
redis = ["redis>=5.0"]
all = [
    "boto3>=1.34",
    "python-jose[cryptography]>=3.3",
    "psycopg[binary]>=3.1",
    "together>=1.0",
    "redis>=5.0",
]

[tool.hatch.build.targets.sdist]
include = ["src/ianuacare", "README.md", "docs"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "src"
packages = ["ianuacare"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["ianuacare"]
branch = true

[tool.coverage.report]
fail_under = 80
omit = ["*/tests/*"]
